Another portion of refactorings
