+++ src/main/java/ru/spbau/mit/VCS.java
@@ -1,0 +1,10 @@
+package ru.spbau.mit;
+
+/**
+ * Created by Эдгар on 04.10.2016.
+ */
+public class VCS {
+    public static VCS getInstance() {
+        return null;
+    }
+}
