--- src/test/java/ru/spbau/mit/core/repository/RepositoryTest.java
@@ -1,11 +1,0 @@
-package ru.spbau.mit.core.repository;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Created by Эдгар on 27.09.2016.
- */
-public class RepositoryTest {
-
-
-}
